tg_api_adapter/requirements.txt
pyTelegramBotAPI
grpcio
protobuf
apscheduler
